third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFECompositeElement-svgdom-k3-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'k3' property of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Math.round(overComposite1.k3.baseVal * 1000) is 2900
PASS Math.round(overComposite1.k3.baseVal * 1000) is 100
PASS successfullyParsed is true

TEST COMPLETE
